Move VERSION to separate file
