+++ lib/tryouts.rb
@@ -4,20 +4,7 @@ unless defined?(TRYOUTS_LIB_HOME)
   TRYOUTS_LIB_HOME = File.expand_path File.dirname(__FILE__)
 end
 
-class Tryouts
-  module VERSION
-    def self.to_s
-      load_config
-      [@version[:MAJOR], @version[:MINOR], @version[:PATCH]].join('.')
-    end
-    alias_method :inspect, :to_s
-    def self.load_config
-      require 'yaml'
-      @version ||= YAML.load_file(File.join(TRYOUTS_LIB_HOME, '..', 'VERSION.yml'))
-    end
-  end
-end
-
+require_relative 'tryouts/version'
 
 class Tryouts
   @debug = false
